+++ CSES/Minimizing Coins.cpp
@@ -13,7 +13,7 @@ void kush_gupta(){
 	#endif
 }
 
-ll dp(vector<ll> &a,ll n,ll x){
+int dp(vector<int> &a,int n,int x){
 	// if (x==0){
 	// 	return 0;
 	// }
@@ -28,16 +28,16 @@ ll dp(vector<ll> &a,ll n,ll x){
 	// 	return dp(a,n-1,x);
 	// }
 
-	vector <vector<ll> >t(n+1,vector <ll> (x+1,0));
-	// ll t[n+1][x+1];
-	for (ll i=1;i<=n;i++){
+	vector <vector<int> >t(n+1,vector <int> (x+1,0));
+	// int t[n+1][x+1];
+	for (int i=1;i<=n;i++){
 		t[i][0]=0;
 	}
-	for (ll j=0;j<=x;j++){
+	for (int j=0;j<=x;j++){
 		t[0][j]=INT_MAX-1;
 	}
-	for (ll i=1;i<=n;i++){
-		for (ll j=1;j<=x;j++){
+	for (int i=1;i<=n;i++){
+		for (int j=1;j<=x;j++){
 			if (a[i-1]<=j){
 				t[i][j]=min(1+t[i][j-a[i-1]],t[i-1][j]);
 			}
@@ -54,10 +54,10 @@ ll dp(vector<ll> &a,ll n,ll x){
 
 int main(){
 	kush_gupta();
-	ll n=0,x=0;
+	int n=0,x=0;
 	cin>>n>>x;
-	std::vector<ll> a(n);
-	for (ll i=0;i<n;i++){
+	std::vector<int> a(n);
+	for (int i=0;i<n;i++){
 		cin>>a[i];
 		// cout<<a[i]<<" ";
 	}
